+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Fixed: Setting/updating a store to a new primitive value, then setting/updating the store to the same value from its subscriptions would prevent `reflect` from being called
 - Fixed: Setting/updating a subscriptionless store to the primitive value it already had could cause a spurious `reflect` call
 
 ## 2.0.0 (June 13, 2019)